+++ src/molecules/DzFormModal/DzFormModal.tsx
@@ -11,7 +11,7 @@ import {
 import { inquireFormSteps } from './formSteps/inquireFormSteps';
 import { newsletterFormSteps } from './formSteps/newsletterFormSteps';
 
-export interface DzFormModalProps {
+export type DzFormModalProps = {
   isOpen: boolean;
   onClose: () => void;
   title: string;
@@ -25,12 +25,15 @@ export interface DzFormModalProps {
   recaptchaNode?: ReactNode;
   disableBackdrop?: boolean;
   onFocus?: Function;
-}
+  onChange?: (fieldName: string, value: any) => void;
+  onDirty?: () => void;
+  LinkElement: any;
+};
 
-export interface SubmissionResult {
+export type SubmissionResult = {
   isSuccess: boolean;
   error?: string;
-}
+};
 
 const FORM_TYPES_TO_STEPS = {
   [FORM_MODAL_TYPES.INQUIRE]: inquireFormSteps,
@@ -51,6 +54,9 @@ export const DzFormModal = ({
   type,
   disableBackdrop = false,
   onFocus,
+  onChange,
+  onDirty,
+  LinkElement = 'a',
 }: DzFormModalProps) => {
   const formSteps = FORM_TYPES_TO_STEPS[type];
   const [submittedFormValues, setSubmittedFormValues] = useState<
@@ -79,7 +85,6 @@ export const DzFormModal = ({
     setIsSubmitSuccessful(undefined);
     onSubmit(formValues)
       .then((result: SubmissionResult) => {
-        setIsSubmitSuccessful(true);
         if (result.isSuccess) {
           setIsSubmitSuccessful(true);
         } else {
@@ -118,9 +123,12 @@ export const DzFormModal = ({
       }
     >
       <DzForm
+        LinkElement={LinkElement}
         steps={formSteps}
         onSubmit={onSubmitForm}
         onFocus={onFocus}
+        onChange={onChange}
+        onDirty={onDirty}
         showStepsCount={false}
         recaptchaNode={recaptchaNode}
         containerClassName="bg-white-100 max-w-[984px]"