+++ src/molecules/DzFormModal/useDzFormModal.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import DzFormModal from './DzFormModal';
 import { FORM_MODAL_TYPE_NAMES } from './types/DzFormModalTypes';
 
-interface UseDzFormModalProps {
+type UseDzFormModalProps = {
   formType: typeof FORM_MODAL_TYPE_NAMES[number];
   onSubmit: (formValues: any) => Promise<any>;
   title: string;
@@ -14,7 +14,10 @@ interface UseDzFormModalProps {
   errorSubtitle?: string;
   disableBackdrop?: boolean;
   recaptchaNode?: ReactNode;
-}
+  onChange?: (fieldName: string, value: string) => void;
+  onDirty?: () => void;
+  LinkElement: any;
+};
 export const useDzFormModal = ({
   formType,
   onSubmit,
@@ -26,6 +29,9 @@ export const useDzFormModal = ({
   errorSubtitle,
   disableBackdrop,
   recaptchaNode,
+  onChange,
+  onDirty,
+  LinkElement = 'a',
 }: UseDzFormModalProps) => {
   const [isOpen, setIsOpen] = useState(disableBackdrop || false);
   const openClickHandler = () => setIsOpen(true);
@@ -45,6 +51,9 @@ export const useDzFormModal = ({
       onSubmit={onSubmit}
       disableBackdrop={disableBackdrop}
       recaptchaNode={recaptchaNode}
+      onChange={onChange}
+      onDirty={onDirty}
+      LinkElement={LinkElement}
     />
   );