+++ CHANGELOG.md
@@ -1,3 +1,87 @@
+# v0.0.140 (Thu Oct 26 2023)
+
+#### 🐛 Bug Fix
+
+- NWEB-705: add split hide media option [#207](https://github.com/Zwirner/design-system/pull/207) ([@JoanCTO](https://github.com/JoanCTO))
+
+#### Authors: 1
+
+- David T ([@JoanCTO](https://github.com/JoanCTO))
+
+---
+
+# v0.0.139 (Thu Oct 26 2023)
+
+#### 🐛 Bug Fix
+
+- NWEB-135 part DzMedia, and some unused components [#205](https://github.com/Zwirner/design-system/pull/205) ([@MadaShindeInai](https://github.com/MadaShindeInai))
+
+#### Authors: 1
+
+- Sergei ([@MadaShindeInai](https://github.com/MadaShindeInai))
+
+---
+
+# v0.0.138 (Wed Oct 25 2023)
+
+#### 🐛 Bug Fix
+
+- rm unintended line [#206](https://github.com/Zwirner/design-system/pull/206) ([@inconduit](https://github.com/inconduit))
+
+#### Authors: 1
+
+- adam ([@inconduit](https://github.com/inconduit))
+
+---
+
+# v0.0.137 (Tue Oct 24 2023)
+
+#### 🐛 Bug Fix
+
+- NWEB-135: DZLink component update. DzHeader support for Link element [#199](https://github.com/Zwirner/design-system/pull/199) ([@MadaShindeInai](https://github.com/MadaShindeInai))
+
+#### Authors: 1
+
+- Sergei ([@MadaShindeInai](https://github.com/MadaShindeInai))
+
+---
+
+# v0.0.136 (Mon Oct 23 2023)
+
+#### 🐛 Bug Fix
+
+- add onDirty callback to form [#204](https://github.com/Zwirner/design-system/pull/204) ([@inconduit](https://github.com/inconduit))
+
+#### Authors: 1
+
+- adam ([@inconduit](https://github.com/inconduit))
+
+---
+
+# v0.0.135 (Mon Oct 23 2023)
+
+#### 🐛 Bug Fix
+
+- provide onChange callback for form modal [#203](https://github.com/Zwirner/design-system/pull/203) ([@inconduit](https://github.com/inconduit))
+
+#### Authors: 1
+
+- adam ([@inconduit](https://github.com/inconduit))
+
+---
+
+# v0.0.134 (Mon Oct 23 2023)
+
+#### 🐛 Bug Fix
+
+- fix SSR effect [#202](https://github.com/Zwirner/design-system/pull/202) ([@inconduit](https://github.com/inconduit))
+
+#### Authors: 1
+
+- adam ([@inconduit](https://github.com/inconduit))
+
+---
+
 # v0.0.133 (Fri Oct 20 2023)
 
 #### 🐛 Bug Fix